Initial commit of working mjpegview pre-alpha
This program was conceived in early September 2011 after Linus Torvalds
opened a Github account and posted his 'diveclog' software, which showed
me how easy it was to write GTK desktop applications. So, having access to
IP cameras, and being slightly frustrated with ZoneMinder, this
application was born.

As currently committed, mjpegview opens a window with a hard-coded display
of two security cameras. The stream and jpeg decoders are functional, but
there's a bug in the framebuf code, so that's not being used yet. Mainly
committing this to have a save point to revert to.