New buffer handling, superior and understandable
Try to reason out the buffer handling with the help of a small test
program instead of stabbing in the dark till something works. Also leave
the cur pointer on the last byte of interest, and advance it only past
it at the end of a state (to put it in position for the next state).

Use more macros for clarity in the stream read code.

Include timing routine for artificial delays; handy when playing back
captured streams.

Change mutex for darea from mutex for pixmap only, to pixmap for the
entire object. So the object is locked from redrawing while resizing.