mjv_thread: create framerate estimator
The framerate estimator keeps a history of timestamps of incoming
frames seen, and calculates an average FPS value from it. When the
difference between the last seen frame and the current wall time
becomes larger than the average delay between frames, the framerate
estimator uses the difference between the first seen timestamp and
the current wall time as an indicator instead.