Big messy commit: major restructuring
These changes were made on a separate branch over the period of a few
months and represent a complex, interweaving set of changes.

The reliance of the mjv_source object on GLib was removed; this lets us
reuse this block in applications that are not linked against GLib (such
as simple headless applications).

The mjv_config code was split off into an mjv_config_source object: this
object represents the config of one single source, and does not depend
on mjv_config. Mjv_config now only generates a linked list of
mjv_config_source objects. This change lets us define mjv_config_source
objects without pulling in a dependency on libconfig, again useful for
simple, headless applications that open a single source.

The file opening and network opening code was moved from mjv_source to
mjv_config_source; all mjv_source expects is an fd from which to read.
Leave the opening of the fd to the part that actually deals with the
source config. This was done in an ongoing effort to make the grabber
state machine less coupled to the rest of the code.

Add the 'mjvsingle' program, a very simple commandline client that takes
arguments on the command line, opens a source, and grabs the image
stream. (Currently doing nothing interesting with it). This simple
client is intended to keep the code "honest": every module it pulls in
should be as simple and self-contained as possible, and should have no
hidden couplings or dependencies that it does not strictly require. The
code in this commit does not strictly enforce that (mjvsingle still has
a dependency on pthread for example), but we're getting there.